+++ packages/core/lib/src/internal/ops/style_border.dart
@@ -28,13 +28,12 @@ class StyleBorder {
                 tree,
                 WidgetPlaceholder(
                   builder: (context, child) {
-                    Widget? built = tree.build() ?? child;
-                    built = _buildBorder(tree, context, built, border);
+                    final built = _buildBorder(tree, context, child, border);
                     return built ?? child;
                   },
                   debugLabel: '${tree.element.localName}--$kCssBorder',
+                  child: tree.build(),
                 ),
-                alignment: PlaceholderAlignment.baseline,
               ),
             );
         },