+++ packages/core/lib/src/internal/ops/tag_ruby.dart
@@ -47,7 +47,6 @@ extension TagRuby on WidgetFactory {
             ),
             debugLabel: kTagRuby,
           ),
-          alignment: PlaceholderAlignment.baseline,
         ),
       );
     }